+++ index.js
@@ -11,7 +11,8 @@ var extend = require('extend');
 var when = require('when');
 var request = require('request');
 var RetryStrategies = require('./strategies');
-var isErrorResponse = require('./helpers/isErrorResponse')
+var isErrorResponse = require('./helpers/isErrorResponse');
+var shouldRejectErrorResponse = require('./helpers/shouldRejectErrorResponse');
 var _ = require('lodash');
 var { StatusCodeError } = require('request-promise-core/errors');
 
@@ -111,7 +112,7 @@ function Request(url, options, f, retryConfig) {
       return this._reject(err);
     }
 
-    if (options.simple && isErrorResponse(response)) {
+    if (shouldRejectErrorResponse(options) && isErrorResponse(response)) {
       return this._reject(new StatusCodeError(response.statusCode, body, this.options, response));
     }